+++ projects/core/src/occ/adapters/user/occ-user-account.adapter.ts
@@ -2,7 +2,7 @@ import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable, throwError } from 'rxjs';
 import { catchError, map } from 'rxjs/operators';
-import { Title, User } from '../../../model/misc.model';
+import { Title, User, UserSignUp } from '../../../model/misc.model';
 import {
   ConsentTemplate,
   ConsentTemplateList,
@@ -19,7 +19,6 @@ import {
 } from '../../../user/connectors/account/converters';
 import { UserAccountAdapter } from '../../../user/connectors/account/user-account.adapter';
 import { USER_NORMALIZER } from '../../../user/connectors/details/converters';
-import { UserRegisterFormData } from '../../../user/model/user.model';
 import { Occ } from '../../occ-models';
 
 const USER_ENDPOINT = 'users/';
@@ -43,7 +42,7 @@ export class OccUserAccountAdapter implements UserAccountAdapter {
     const endpoint = userId ? `${USER_ENDPOINT}${userId}` : USER_ENDPOINT;
     return this.occEndpoints.getEndpoint(endpoint);
   }
-  register(user: UserRegisterFormData): Observable<User> {
+  register(user: UserSignUp): Observable<User> {
     const url: string = this.getUserEndpoint();
     let headers = new HttpHeaders({
       'Content-Type': 'application/json',
