+++ projects/core/src/user/connectors/account/converters.ts
@@ -1,10 +1,9 @@
 import { InjectionToken } from '@angular/core';
 import { Converter } from '../../../util/converter.service';
-import { UserRegisterFormData } from '../../model/user.model';
-import { Title } from '../../../model/misc.model';
+import { Title, UserSignUp } from '../../../model/misc.model';
 
 export const USER_REGISTER_FORM_SERIALIZER = new InjectionToken<
-  Converter<UserRegisterFormData, any>
+  Converter<UserSignUp, any>
 >('UserRegisterFormSerializer');
 
 export const TITLE_NORMALIZER = new InjectionToken<Converter<any, Title>>(
