+++ projects/core/src/user/connectors/account/user-account.adapter.ts
@@ -1,13 +1,12 @@
 import { Observable } from 'rxjs';
-import { Title, User } from '../../../model/misc.model';
+import { Title, User, UserSignUp } from '../../../model/misc.model';
 import {
   ConsentTemplate,
   ConsentTemplateList,
 } from '../../../occ/occ-models/additional-occ.models';
-import { UserRegisterFormData } from '../../../user/model/user.model';
 
 export abstract class UserAccountAdapter {
-  abstract register(user: UserRegisterFormData): Observable<User>;
+  abstract register(user: UserSignUp): Observable<User>;
 
   abstract requestForgotPasswordEmail(userEmailAddress: string): Observable<{}>;
