+++ projects/core/src/user/connectors/account/user-account.connector.ts
@@ -1,11 +1,10 @@
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { Title, User } from '../../../model/misc.model';
+import { Title, User, UserSignUp } from '../../../model/misc.model';
 import {
   ConsentTemplate,
   ConsentTemplateList,
 } from '../../../occ/occ-models/additional-occ.models';
-import { UserRegisterFormData } from '../../../user/model/user.model';
 import { UserAccountAdapter } from './user-account.adapter';
 
 @Injectable({
@@ -14,7 +13,7 @@ import { UserAccountAdapter } from './user-account.adapter';
 export class UserAccountConnector {
   constructor(protected adapter: UserAccountAdapter) {}
 
-  register(user: UserRegisterFormData): Observable<User> {
+  register(user: UserSignUp): Observable<User> {
     return this.adapter.register(user);
   }
