+++ projects/core/src/user/index.ts
@@ -2,7 +2,6 @@ export * from './store/actions/index';
 export * from './store/selectors/index';
 export * from './store/user-state';
 
-export * from './model/user.model';
 export * from './facade/index';
 export * from './user.module';
 export * from './connectors/index';
