--- projects/core/src/user/model/user.model.ts
@@ -1,7 +1,0 @@
-export interface UserRegisterFormData {
-  lastName: string;
-  firstName: string;
-  uid: string;
-  password: string;
-  titleCode: string;
-}
