+++ projects/core/src/user/store/effects/user-register.effect.ts
@@ -7,8 +7,8 @@ import { catchError, map, mergeMap, switchMap } from 'rxjs/operators';
 
 import * as fromActions from '../actions/user-register.action';
 import { LoadUserToken, Logout } from '../../../auth/index';
-import { UserRegisterFormData } from '../../../user/model/user.model';
 import { UserAccountConnector } from '../../connectors/account/user-account.connector';
+import { UserSignUp } from '../../../model/misc.model';
 
 @Injectable()
 export class UserRegisterEffects {
@@ -18,7 +18,7 @@ export class UserRegisterEffects {
   > = this.actions$.pipe(
     ofType(fromActions.REGISTER_USER),
     map((action: fromActions.RegisterUser) => action.payload),
-    mergeMap((user: UserRegisterFormData) => {
+    mergeMap((user: UserSignUp) => {
       return this.userAccountConnector.register(user).pipe(
         switchMap(_result => [
           new LoadUserToken({
